+++ src/Token.php
@@ -12,11 +12,16 @@ use function preg_match;
  */
 final class Token
 {
-    private function __construct(public readonly string $value)
+    private function __construct(private readonly string $value)
     {
         if (1 !== preg_match("/^([a-z*][a-z\d:\/!#\$%&'*+\-.^_`|~]*)$/i", $this->value)) {
             throw new SyntaxError('Invalid characters in token.');
         }
+    }
+
+    public function toString(): string
+    {
+        return $this->value;
     }
 
     public static function fromString(Stringable|string $value): self